(chore) add support for the 'order' property when connector/action form config specifies it
